Printer-friendlier layout, esp. for ranking
Adapted the bootstrap/html5boilerplate css to tradity.
Only the main content is visible in print mode and ranking UI
like the selector buttons are not shown.
Also, the group sponsors are omitted from the printout.
A special print button is added to the ranking, ensuring that all
entries are shown before opening the browser’s printing dialogue.