skip log in if logged in
If you are already logged in and go to the log-in-page you have to log in again, but you should actually directly go to the dashboard. This is fixed by checking the "authToken" if it's null. It works fine but I am not sure if I used the constructor correctly by adding the "apiService" to it.